scorbot: use differential kins for the wrist
The Scorbot-ER 3 has a differential-drive wrist with two motors and two
limit switches.  The limit switches sense the limits of wrist roll and
wrist pitch, but each motor drives both roll and pitch.  The two wrist
motors need to be moved in a coordinated way to home the wrist.

This is problematic since Motion homes joints independently of each other.

This commit adds A and B axes to the scorbot kinematics, representing
wrist roll & pitch, passed straight through to HAL.

In HAL, the new "differential" comp handles converting wrist roll and
pitch commands from Motion into coordinated motor position commands for
the J3 and J4 motors.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>