Gantry improvements in saved positions
We now find the primary joint for an axis with more than one joint,
and save the position for that joint in all the others on the axis.
This fixes the problem of a following error after a failed homing or such
where you have to go delete the position.txt file in order to even run again.