New custom homing: candcnchoming.comp
This custom homing module handles gantries the way they should have been to start with.
Each move is synchronized instead of only the final move to home position.
Previously each side of the gantry could be moving in opposite directions at the same time,
resulting in severe racking if the home switches had a long throw.
This fixes that so each side stops when it hits it's target and waits for the other joint before
starting the next move.